// src/api/openapi/models/language.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Text } from './text';


export interface Language { 
    code: string;
    id: number;
    name: Text;
}