// src/api/openapi/models/clientCreate.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AddressCreate } from './addressCreate';
import { ContactCreate } from './contactCreate';


export interface ClientCreate { 
    address: AddressCreate;
    cig: string;
    codice_destinatario: string;
    contact_person: string;
    contacts: Array<ContactCreate>;
    cup: string;
    esigibilita_iva: string;
    fiscal_code: string;
    gender_code: string;
    isCompany: boolean;
    language_code: string;
    lastname: string;
    mail1: string;
    mail2: string;
    name: string;
    pec: string;
    publica_amministrazione: boolean;
    tel1: string;
    tel2: string;
    vat_number: string;
}